In the shuttle Task, I looped through all the numbers from 1 to 100 using a for loop. I checked each number to see if it was an unlucky shuttle number by checking its last digit (i % 10) and first digit (i / 10), and if they are both equal to 4 or 9, then we skip that number using 'continue' operator. If the number is not an unlucky shuttle number, then we print it on the screen using the println method.
Number task
Created a variable called secretNumber.
Created a Random class.
Created a Scanner class for reading input from the console.
Using the nextInt method of the Random class, generate a random number in the range from 0 to 10 and save it in the secretNumber variable.
Display a message with hints for the user.
The while loop checks whether the user has exceeded the number of attempts (3). If so, we display a message about the loss and terminate the program.
Request the user's number and save it in the guess variable.
If the user's number matches the secret number, we display a message about the win and terminate the program.
If the user's number does not match the secret number, we display a message that they did not guess and continue the loop.
If the user did not guess the number in 3 attempts, we display a message about the loss and tell them what number the computer had chosen.